ENH: use telescope and dut objects throughout the analysis, major cleanup and brushup

+++ testbeam_analysis/cpp/data_struct.pxd
@@ -3,14 +3,14 @@ cnp.import_array()  # if array is used it has to be imported, otherwise possible
 
 cdef packed struct numpy_hit_info:
     cnp.int64_t eventNumber  # event number value (unsigned long long: 0 to 18,446,744,073,709,551,615)
-    cnp.uint8_t frame  # relative BCID value (unsigned char: 0 to 255)
+    cnp.uint32_t frame  # relative BCID value (unsigned char: 0 to 255)
     cnp.uint16_t column  # column value (unsigned char: 0 to 255)
     cnp.uint16_t row  # row value (unsigned short int: 0 to 65.535)
     cnp.uint16_t charge  # ToT value (unsigned char: 0 to 255)
 
 cdef packed struct numpy_cluster_hit_info:
     cnp.int64_t eventNumber  # event number value (unsigned long long: 0 to 18,446,744,073,709,551,615)
-    cnp.uint8_t frame  # relative BCID value (unsigned char: 0 to 255)
+    cnp.uint32_t frame  # relative BCID value (unsigned char: 0 to 255)
     cnp.uint16_t column  # column value (unsigned char: 0 to 255)
     cnp.uint16_t row  # row value (unsigned short int: 0 to 65.535)
     cnp.uint16_t charge  # ToT value (unsigned char: 0 to 255)
